Fix crash in call log.
The latest CoordinatorLayout code checks for being used with an
AppCompat theme on creation. As we neither need nor want to go the full
AppCompat route and the CoordinatorLayout works fine without the
AppCompat theme for what we're using it for, pretend our theme is
AppCompat based just to make the check happy.

Change-Id: I5fa321790fe539ceee652da8416297e009e1c495